construct edges for schema concepts in the visualiser
